net-misc/curl: add 8.5.0
This release includes fixes for two CVEs linked as bugs below.

For clarity, Gentoo does not currently enable a libpsl dependency so
version bump does not address CVE-2023-46218 / Bug 919325; we're
no more or less vulnerable than we were before.

Test 1477 has been disabled for this release; it's docs related
and upstream did not include a required file - it's not worth us
patching:
https://github.com/curl/curl/commit/da8c1d15782c8161b455a7ee90197c16ae5edb90

Bug: https://bugs.gentoo.org/919325
Bug: https://bugs.gentoo.org/919326
Signed-off-by: Matt Jolly <Matt.Jolly@footclan.ninja>