dev-php/theseer-Autoload: Do not symlink outside of EPREFIX
During the prepare step of the ebuild, symlinks are created to
previously installed PHP dependencies in the vendor subdirectory to
bootstap phpab. These symlinks lead to the real root filesystem in
/usr/share/php/... instead of the portage prefix $EPREFIX/usr/share/php,
thus being essentially broken in a Prefix installation.

By adding $EPREFIX to the link directory, the link is now correct
both inside and outside of Prefix installations.

Closes: https://bugs.gentoo.org/918933
Signed-off-by: Moritz Brunner <moritz.brunner+gentoo@posteo.de>
Signed-off-by: Fabian Groffen <grobian@gentoo.org>