profiles: new LLVM_PROJECTS USE_EXPAND
This USE_EXPAND flag controls the projects that will be built in an LLVM
monorepo build. It's intended to allow users to build LLVM without
needing to build the entirety of the projects contained within LLVM. It
should follow the values that LLVM accepts to its LLVM_ENABLE_PROJECTS
CMake option.

Signed-off-by: Violet Purcell <vimproved@inventati.org>