media-video/ffmpeg: Conditionally apply patch set for playback on SoCs
As much as we hate deviating from upstream, this patch set is really
important for smooth playback on SoCs commonly used for ARM and RISC-V,
including the Raspberry Pi. We will enable the "soc" USE flag by default
on these architectures.

These are different patch sets floating around, but this one from John
Cox of Kynesim is the most comprehensive. It is the only one that
enables smooth playback for the Pi under Kodi, while benefiting other
systems at the same time. John keeps the set reasonably in sync with
upstream. The road is long, but the hope is that the changes will
eventually be upstreamed.

Note that LibreELEC only applies this patch set specifically for its Pi
builds, while using lighter sets for some other builds, but they have
told me that if we're only going to use one set then this is the one to
have.

Following this, also drop legacy Raspberry Pi support as
media-libs/raspberrypi-userland is being last-rited.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>