profiles/use.desc: Make USE=gles2 global
Add a global USE=gles2 flag, to enable GLES 2 support (as a companion
to USE=gles2-only).  It seems to be used consistently in 16 packages,
though for some it also enables GLES 3).

Technically there are also `gles1` and `gles3` flags in the wild
but they are used by a few packages.

Signed-off-by: Michał Górny <mgorny@gentoo.org>