dev-libs/libtomcrypt: Fix cross-compiling by using libtool from SYSROOT
libtool is normally generated by configure with the right toolchain
settings, but this package does not use autoconf, so it executes the
build host libtool with the wrong settings. CC seems to be respected for
compiling but not linking for some reason.

As much as I hate executing something from SYSROOT, hell will freeze
over before libtool stops being a shell script. This seems far easier
than any alternative. ${BASH} at least ensures that it is not executed
using some prefixed shebang.

We'll still allow ${LIBTOOL} to be manually overridden because slibtool
does not suffer from this issue.

Closes: https://github.com/gentoo/gentoo/pull/34334
Signed-off-by: James Le Cuirot <chewi@gentoo.org>