dev-python/pypy3_9: fixup compileall call excludes
There's a bunch of testdata that isn't supposed to be compiled. The
CPython build system automatically excludes it, but PyPy doesn't have a
build system so we must do it manually.

The original implementation passes the excludes args after the operand,
which is unreliable across PyPy versions. For consistency, specify them
before.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>