dev-qt/qttools: disable clang test with USE=-clang
Normally harmless (tries to find it, either does or not and will not
use it either way), but if the cmake files are in broken state then
it can abort entirely.

Unsure if it fully resolves bug #916098 (for portage, ideally these
should still be updated "together" as much as possible), but at least
should not trigger on qttools anymore for those not enabling clang
(tested by intentionally breaking llvm's cmake files).

Closes: https://bugs.gentoo.org/916098
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>