dev-util/perf: mirror kernel restrictions for libdw support
Kernel has a whitelist of architectures for which unwind via libdw is
supported.  Right now the ebuild unconditionally enables it, this fails
on architectures where kernel support is not merged (e.g. alpha).

This list needs to be checked and synced on updates.  There shouldn't be
anything removed from it in the future, only added.

Bug: https://bugs.gentoo.org/759754
Signed-off-by: Matoro Mahri <matoro_gentoo@matoro.tk>