dev-util/source-highlight: remove unneeded runtime dependency on ctags
ctags is not actually needed to highlight files. It is used at runtime
for --gen-references, which isn't the default way to use it, and
source-highlight knows how to error out with a friendly error message if
you do not have it and need it.

It's used by the testsuite, of course.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>