dev-util/meson: solve the python2 test breakage correctly this time
The problem is that python-utils-r1 adds some wrappers around all the
python impls. I am not sure what these wrappers do (we don't even use
the python3 one, since the PEP517 installed copy of meson comes before
that in PATH, and overrides it) but it also poisons the `python2`
command with a script that says it is unsupported.

Well, no kidding it's unsupported to invoke building and installing a
python3.11 wheel using a `python2` command. However, we are not actually
trying to do this. So we should not feel constrained by this, either.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>