dev-cpp/simpleini: add 4.21
CMakeLists.txt is now more usable (minus some quirks) and should
probably be switched to in the future. But it changes the install
location of the header and its only revdep (devilutionx) currently
cannot find it. Could be solved with a 1 line sed/patch in devilutionx
but will wait and stick to makefile for now to observe the situation.

Will reconsider next simpleini and/or devilutionx bump.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>