www-servers/apache: clean up ancient outdated split-usr hack
For nearly a decade, USE=split-usr has been declared in apache-2.eclass
and provides a hack intended as a short-term migration for moving a
binary from /usr/sbin to /usr/bin.

No remaining cases of /usr/sbin/apxs exist in tree, so drop the symlink
and the USE flag and revbump.

Closes: https://bugs.gentoo.org/502384
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>