sys-power/nut: fix confusing split-usr handling
Almost a decade ago, upstream moved their binary from libdir to sbindir.
Historically, Gentoo symlinked the binary into sbindir to make up for
this -- and when upstream moved to follow the Gentoo layout, the Gentoo
symlink overwrote the actual binary.

For some reason, at the time, the solution used was to add a
USE=split-usr and install extra symlinks for /sbin -> /usr/sbin even
though this didn't actually do something meaningful. The split-usr USE
flag has issues and using it unnecessarily should be avoided if
possible. Drop the USE flag, and the symlink, and stick with /usr/sbin
as upstream intends.

This should be 100% fine in all cases. The /sbin directory is only
needed for users who both have /usr on a different partition, *and*
don't use an initramfs. In such a case, it is nonsensical to provide a
symlink in /sbin that is a broken symlink to an un-mounted partition.

Bug: https://bugs.gentoo.org/517954
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>