sys-cluster/drbd-utils: avoid the necessity of USE=split-usr
It was added in commit 02512e1d120272ad4e035959fa4014a8df7a28fc without
any explanation. Comparing builds, it does matter, though -- because on
split-usr systems, it uses and needs the udevdir to be /lib/udev.

Upstream detects this via pkg-config automagic with no way to override
it. But the ebuild never actually depended on virtual/pkgconfig
anyway... the good news is it only uses it for two things.

The other thing it uses it for is for detecting systemd to set the
default systemd unit directory when the configure option is not passed.
We want to install it unconditionally though, which means we cannot rely
on pkg-config for this either. That means we are using it for nothing.

Instead, explicitly set PKG_CONFIG=/bin/false to ensure it gets detected
as missing/broken, which means we fall back to the default udev
directory value, which is safe for split-usr systems.

While we are at it:

- ensure the systemdunitdir setting actually does something by declaring
  the init type as systemd. The openrc files come from gentoo anyway and
  are overwritten after emake install.
- avoid reinstalling the tmpfiles, which was originally added in commit
  85d6acdce7c5e60c5dfe768d1b3514e2c962c2ea specifically to work around
  the "only on systemd" issue.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>