net-fs/davfs2: remove pointless no-op USE=split-usr
In bug 884323 it was noted that merged-usr systems get a conflict,
because upstream attempted to install a symlink in /sbin pointing to
/usr/sbin. It was noted in the bug that /sbin isn't needed at all,
despite which the chosen solution was to add IUSE=split-usr and
conditionally create the (useless) symlink.

In version 1.7.0 upstream dropped this entirely, so no symlink is
created either way -- but we still defined the USE flag and invalidated
binpkg changed-use matches. Also we wasted ebuild real estate redefining
the src_install function.

Get rid of this dead weight.

Bug: https://bugs.gentoo.org/884323
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>