sys-apps/iproute2: drop old compat symlink and USE=split-usr hack
As promised in the ebuild comments, we can now remove the compatibility
symlink from /sbin -> /bin.

While we are at it, don't do the move conditional on split-usr. Arguably
only the symlink should have been conditionalized on this way back
when... we already move another binary too in the same way, but without
any conditional.

Bug: https://bugs.gentoo.org/547264
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>