media-sound/guitarix: restore parallel build
Build system messes with the build jobs count, so that waf build
always respects the number of jobs passed to waf configure. And
since waf-utils eclass passes --jobs=1 to waf configure, we are
always building using only 1 job.

Pass --jobs with the correct number of jobs to waf configure as a
workaround.

Closes: https://bugs.gentoo.org/908690
Signed-off-by: Alexander Tsoy <alexander@tsoy.me>