dev-qt/qtwebengine: update patchset for 6.6 and 6.7 branches
6.6-5 -> 6.6-6 changes (only used by 6.6.1):
  Added:
  * gcc14.patch (imported from gentoo tree, not needed in 6.7)
  Updated:
  * cstdint.patch (>=qtwebengine-6.6.1 needs an extra for gcc+musl)

6.6-6 -> 6.6-7 changes (due for 6.6.2 if nothing else comes up):
  Removed:
  * libcxx17.patch (backported from 6.7)
  * libxml2-2.12.patch (upstreamed)

6.7-1 -> 6.7-2 changes (tentative, release is still far away):
  Updated:
  * cstdint.patch (>=qtwebengine-6.6.1 needs an extra for gcc+musl)
  Removed:
  * libxml2-2.12.patch (upstreamed)

Safe changes, and no need for revbumps.

For libxml2, Qt did the same changes that we did rather than do
like upstream chromium. Meaning we do not need to depend on the
newer libxml2 and it works for the old as well.

Hopefully more gcc+musl issues don't keep showing up given been using
clang+musl for testing musl and would rather not do both on top of
gcc+glibc (idea was to pickup most toolchain issues with only two
builds). Only know about this because a user mentioned it on IRC.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>