dev-qt/qtwebengine: do the install sanity-check from qt5 in qt6 too
The check was removed before I touched these and didn't know it
existed. Was fair to think it wouldn't be needed anymore with cmake
but it is, e.g. it installs nothing if bison is not found without
hard failure.

File path is hopefully correct for Qt6, did give it two test builds
(one "bad" that's empty, and one full build) just in case. Albeit
haven't retried prefix (it's included in QT6_LIBDIR though).

The qt6-build eclass does force fatal errors for a few "build nothing"
cases, but not qtwebengine's custom ones. Kind of wonder if a similar
end result file check should be done for all of dev-qt/* (aka fail if
does not install anything but cmake files and docs). May revisit if Qt
has more original ways to build nothing not limited to qtwebengine.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>