x11-drivers/nvidia-drivers: fix with linux-6.7 and some configs
Looked fine for me but turns out the issue is conditional to kernel
configs and it was not enabling USE_LKCA with mines. Also needed
USE=kernel-open to run into this.

An NVIDIA representative has offered a fix, so may as well use it
this time around.

525 and older branches did not use this function.

Should be harmless for older kernels, but have not really tested
so only applying if >6.7 as a precaution, esp. for stable 535.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>