media-gfx/sam2p: Fix clang -flto build failure
When compiling with clang and ld.lld build fails with the error message:

    ps_tiny.o: file not recognized: file format not recognized

This is because the build system does not properly honor standard
buildtime flags and instead requires nonstandard flags be set to get the
proper build result. This patch sets the necessary flags to successfully
compile with clang and -flto.

Signed-off-by: Nicholas Vinson <nvinson234@gmail.com>