net-libs/pjproject: bump to 2.13.1-r1 to add IUSE=srtp
Bump to make linking in libsrtp optional so we can keyword asterisk for
sparc.

There are no known project in gentoo other than asterisk, which don't
care for pjproject having (or not having) srtp.  Thus in order to enable
pjproject to be used on sparc we need to be able to not depend on libsrtp.

Bug: https://bugs.gentoo.org/919424
Signed-off-by: Jaco Kroon <jaco@uls.co.za>