net-proxy/tayga: Fix "incompatible pointer to integer conversion"
The purpose of the code is to find the separator slash of the CIDR
range, splitting it to the IP and the number of bits in the range.
Assigning NUL character to the memory location completes the string
split, but the code assigns NULL pointer instead. While functionally
the same, the compiler type checking complains.

I'm not revbumping this change because the code should compile the
same in theory.

Closes: https://bugs.gentoo.org/878629
Signed-off-by: YiFei Zhu <zhuyifei1999@gmail.com>
Signed-off-by: Michał Górny <mgorny@gentoo.org>