app-emulation/wine-proton: make libdrm[video_cards_amdgpu] optional in live
Had forgotten about it but wine offers generic switches for all dlls,
so can use that rather than modify sources. The dll being entirely
missing hopefully has no negative impact (should gracefully just not
use it even for amdgpu).

Imagine may be new releases soon'ish (or proton 9), and meant to adjust
this before then as it'll be annoying for non-amdgpu users to enable it
on libdrm.

May possibly need masking on x86 (32bit), have not tested if it builds
anyway. Albeit doubt this is usable there regardless. Generally unsure
if the proton variant should even keep x86 keywords though (was kind
of flaky when I tested it for x86 while considering stable and not
entirely sure it works on a real system).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>