net-dns/libidn: add 1.42
Note that I've gone back to the default src_test implementation. Ignoring
gnulib test failures isn't ideal at all, so let's not do that, given it could
affect how the program operates and we've found other bugs that way in other
packages before. I'm not aware of any other packages where we make an effort
to skip the gnulib tests.

Bug: https://bugs.gentoo.org/500662
Bug: https://bugs.gentoo.org/539356
Closes: https://bugs.gentoo.org/898348
Signed-off-by: Sam James <sam@gentoo.org>