games-emulation/pcsx2: stabilize 1.7.5312 for amd64
1.6.0 is the current "stable" release upstream, but is:
* nearly 4 years old (lacks a *lot* of game fixes)
* 32bit-only (needs multilib on amd64)
* starting to hit issues like bug #922149 (which took a long
  time to be reported, indicating it likely has rather few users)
* has some utf8 locale issues last time I tried it
* ebuild is bit dated and needs update for newer wxgtk
* not really getting used/tested by its gentoo maintainers
  with no real motivation to fix problems

The 64bit Qt6 1.7.x may still be kind of WIP but has matured a fair
amount and 1.6.0 does not seem to offer much advantages over it with
the exception of working on older machines (x86, or anything lacking
sse4.1).

An upstream "real" release will likely take a long time still, and
the situation for 1.6.0 just get worse meanwhile.

So, opting to stable 1.7.x even though it's "nightly" arbitrary
releases and drop 1.6.0 given it's broken in current state.

Bug: https://bugs.gentoo.org/922149
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>