games-emulation/pcsx2: update live, require clang
Likely an easy fix that could be done from our end, but no
longer have the patience to support things that upstream is
refusing to support. Albeit may drop the requirement if it
gets fixed by the time of next bump, but otherwise it'll
likely stick around.

Likely not worth using the llvm eclass here, not linking with
and only need to pick the current compiler.

fwiw forcing clang removes the need for the lto patch.

Also BUILD_DATE option was removed. GIT_TAG bits may also need
updates but will check that next time do a bump as it does not
matter in live. LICENSE may need a bit of reviewing too given
upstream added a compilation of all the licenses used.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>