sys-kernel/dracut: add 060_pre20240104
Add patch to explicitly support grub layout in 50-dracut.install

Ensures that the generated initrd is installed in the "new" way via the
staging area in the grub layout. This prevents accidentally creating
an UKI named initrd, and also ensures that BOOT_ROOT and
KERNEL_INSTALL_INITRD_GENERATOR are respected when the layout is set to grub.

Downstream only since the grub layout for using grub with systemd's
kernel-install is not supported by systemd upstream and therefore this
patch is unlikely to be accepted by dracut upstream.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>