x11-drivers/nvidia-drivers: fix build with upcoming gcc14
Trivial and proper fix for 5xx branches, NVIDIA seems to be
(now) keeping track of implicits and incompatibles beside
missing this one hidden behind 2>/dev/null and kernel >=5.18.

Can't say the same for the legacy branches (390 and 470), and
rather than worry about these going for the lame life support
treatment with -Wno-error= there. 470 has hope to be fixed
properly by NVIDIA still but this is likely there forever in
the not supported + masked 390.

No need for revbumps, with gcc13 the test simply functions
properly and does what's right, while with 14 it's just a
build-time issue (and legacy branches are unchanged).

Closes: https://bugs.gentoo.org/921370
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>