dev-util/mingw64-toolchain: prevent makeinfo usage
This does not install .info pages (that is left to binutils+gcc
real packages), so little reason to depend on texinfo(makeinfo)
and generate them for nothing.

Could possibly come back as an issue if use gcc snapshots (that
lack pre-gen info pages), and unsure if this same workaround will
work for gcc (currently only binutils is trying to use it).

MAKEINFO=: on ./configure is not necessary (and is insufficient
to stop usage given it is not respected in Makefiles), but it
prevents a command not found QA notice.

That aside, generally hard for texinfo to be missing unless
depclean build deps, so BDEPEND would not have been too wasteful.

Closes: https://bugs.gentoo.org/922230
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>