Revert "dev-util/meson-1.3.1: fix for Darwin with native linker again"
This reverts commit b7035fb0da8ffcf1577b68d43f49511adee8237d.

This patch was previously introduced for bug 868516, and hit pushback by
multiple parties that wanted to see this discussed upstream. After some
coaxing and arm-twisting, an upstream issue was finally opened (but no
patch submitted). The patch in ::gentoo went stale, and got dropped.
base-system@ is uninterested in maintaining this out of tree patch given
the situation (and neither am I).

After the ticket was opened upstream, it was retracted by the submitter:

> I decided to fix the problem from out custom version of xtools's end.

Now it's back as a local patch to dev-build/meson, where it's just going
to bitrot another time? No explanation for why this is necessary,
especially if xtools added compatible output. No attempt to submit a PR
to meson. No bug link to cross-reference relevant bugs.

Solve this by reducing the local-patches tech debt and punting on the
issue, pending actual upstreaming. We can revisit backporting a patch if
and when it will constitute a backport of a patch available in upstream
git master.

Bug: https://bugs.gentoo.org/868516
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>