app-emulation/libvirt: Update live ebuild
Starting from its upstream commit of v9.9.0-rc1~108 libvirt can
now use nbdkit to access network disks. The commit introduced new
meson option (-Dnbdkit), which is basically just a rename of an
old one (-Dlibnbd). We didn't expose the old one, but it makes
sense to export the new one for proper deps tracking.

But because nbdkit is available only on a small subset of arches,
the useflag has to be masked on other arches that doesn't have
sys-block/nbdkit.

Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>