app-emulation/wine-vanilla: mention toolchain in wow64 description
Or else it may come as confusing if users try to disable
USE=abi_x86_32 and then a dependency still requests it.

mingw64-toolchain is the only one needed, no need for glibc, gcc,
mesa/nvidia, gstreamer, etc... and it works even on no-multilib
profiles where the USE is unmasked.

There are alternatives to mingw64-toolchain, but let's not go
into confusing details in the USE desc.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>