app-arch/cpio: run tests in parallel
Inspired by vapier's cd7f047fe43fb631c4ca6979c6efb5038c616f41. `RUNTESTFLAGS`
has been in the back of my head for a while now as I've been meaning to figure out
the interaction between it and parallel make and which packages actually need it.

This only shaves off about ~10s or so consistently for me for cpio though, as
its test suite is smaller, but maybe it'll do better on slower machines.

Signed-off-by: Sam James <sam@gentoo.org>