sys-devel/bison: run tests in parallel
Inspired by vapier's cd7f047fe43fb631c4ca6979c6efb5038c616f41. `RUNTESTFLAGS`
has been in the back of my head for a while now as I've been meaning to figure out
the interaction between it and parallel make and which packages actually need it.

This saves about 30s for me on a fast machine.

Signed-off-by: Sam James <sam@gentoo.org>