dev-util/lcov: Add patch to fix bad error code var
There is a mistake in the lcov code for inconsistent line numbers, the
ignorable error code is an undefined variable. This causes errors like
genhtml: ERROR: "/mnt/host/source/src/platform/ec/common/vboot/vb21_lib.c":18: function 'vb21_is_packed_key
_valid last line is 24 but is 25 in testcase 'vboot'
        (use "genhtml --ignore-errors code_ ..." to bypass this error)

Switch to the correct variable that was fixed in
https://github.com/linux-test-project/lcov/commit/6f48bf214a06fe09c9708b130a233280c691a7a0#diff-173ac1b8ad02cbcecdc47f8c7fa7762bbcc7050d0efa6a375ffb74eb1517f68cL5342
but not included in the 2.0 release.

Signed-off-by: Jeremy Bettis <jbettis@google.com>